Don't try to run command in obsolete container (locally built).
Before running a command, which should be run in a docker container,
defined by local Dockerfile, rego tries to (re)build those container.
Currently, rego ignores errors, which happen at build step.
For the very first run it will lead to error anyway, because rego
will not be able to run container, which is not built yet.
But if container was already build in the past, later some error(s)
were introduced to Dockerfile and build fails, rego will use old
version of the container, what is incorrect.
It should clearly state that build of container failed.