Rename 'rego' to 'dev-it-easy'/'dev'
It appears to be extreamly hard to find good name, which is not used yet,
but at the same time it is short and meaningful.

Unfortunately 'rego' - it is some policy language, which is already exists.

Which alternative were considered:
- justrun/jr
- dev-only/do
- drun
- erun
- doit
- rcli
- rcmd
- edev

and many many more. I think that current name is not ideal, but at the moment it is a compromise and I didn't find anything better.